mv icp -> traits
